[refactor] Rename MeasureCounts.* -> SampleResult.*
