+++ runtime/cudaq/qis/execution_manager.h
@@ -9,9 +9,9 @@
 #pragma once
 
 #include "common/CustomOp.h"
-#include "common/MeasureCounts.h"
 #include "common/NoiseModel.h"
 #include "common/QuditIdTracker.h"
+#include "common/SampleResult.h"
 #include "cudaq/host_config.h"
 #include "cudaq/operators.h"
 #include <deque>
