+++ unittests/common/NoiseModelTester.cpp
@@ -8,7 +8,7 @@
 
 #include "CUDAQTestUtils.h"
 #include "common/FmtCore.h"
-#include "common/MeasureCounts.h"
+#include "common/SampleResult.h"
 
 using namespace cudaq;
